add dao methods
