+++ src/main/groovy/dao/impl/UserDaoImpl.groovy
@@ -7,12 +7,12 @@ import exception.RecordCouldNotSavedException
 import exception.SamePasswordException
 import exception.UserAlreadyExistsException
 import exception.UserNotExistsException
-import exception.InvalidTokenException
 import groovy.util.logging.Slf4j
 import model.AuthenticationResponse
 import model.ResponseData
 import org.jooq.DSLContext
 import org.jooq.Record
+import org.mindrot.jbcrypt.BCrypt
 import org.restlet.data.MediaType
 import org.restlet.representation.Representation
 import org.restlet.representation.StringRepresentation
@@ -37,7 +37,7 @@ class UserDaoImpl implements UserDao {
 
   @Override
   Representation signUpUser(Userdata userdata) {
-    if (userDataValidation.isUserExists(userdata.getUsername())) {
+    if (isUserExists(userdata.getUsername())) {
       log.error("User already exists with username {}", userdata.getUsername())
       throw new UserAlreadyExistsException("user exists with given username")
     } else {
@@ -55,11 +55,10 @@ class UserDaoImpl implements UserDao {
     }
   }
 
-
   @Override
   AuthenticationResponse signInUser(Userdata userdata) {
     Record userRecord = dslContext.fetchOne(dslContext.selectFrom("UserData").where(USERDATA.USERNAME.eq(userdata.getUsername())))
-    if (!userDataValidation.isUserExists(userdata.getUsername())) {
+    if (!isUserExists(userdata.getUsername())) {
       log.error("User not exists with username {}", userdata.getUsername())
       throw new UserNotExistsException("User not exists with given username: " + userdata.getUsername())
     } else if (!userDataValidation.isPasswordCorrect(userdata, userRecord)) {
@@ -73,11 +72,10 @@ class UserDaoImpl implements UserDao {
 
   @Override
   Representation changeUserPassword(Userdata userdata, int userId) {
-    Record userRecord = dslContext.fetchOne(dslContext.selectFrom("UserData").where(USERDATA.ID.eq(userId)))
-    if (!userDataValidation.isUserExistsById(userId)) {
+    if (!isUserExistsById(userId)) {
       log.error("User not exists with the user userId: {}", userId)
       throw new UserNotExistsException("User not exists with given id:" + userId)
-    } else if (!userDataValidation.isOldPasswordCorrect(userdata.getPassword(), userId)) {
+    } else if (!isOldPasswordCorrect(userdata.getPassword(), userId)) {
       throw new InvalidPasswordException("Password is invalid for given user id" + userId)
     } else if (userDataValidation.isNewPassWordSame(userdata)) {
       throw new SamePasswordException("Try another password ")
@@ -96,12 +94,28 @@ class UserDaoImpl implements UserDao {
 
   @Override
   ResponseData getUserData(int userId) {
-    if (userDataValidation.isUserExistsById(userId)) {
+    if (isUserExistsById(userId)) {
       Record userRecord = dslContext.fetchOne(dslContext.selectFrom("UserData").where(USERDATA.ID.eq(userId)))
       return new ResponseData(userRecord.get("id") as int, userRecord.get("username") as String)
     } else {
       log.error("User not exists with user id {}", userId)
       throw new UserNotExistsException("User not exists with given id:" + userId)
     }
+  }
+
+  @Override
+  boolean isUserExists(String username) {
+    return dslContext.fetchExists(dslContext.selectFrom("UserData").where(USERDATA.USERNAME.eq(username)))
+  }
+
+  @Override
+  boolean isUserExistsById(int id) {
+    return dslContext.fetchExists(dslContext.selectFrom("UserData").where(USERDATA.ID.eq(id)))
+  }
+
+  @Override
+  boolean isOldPasswordCorrect(String password, int id) {
+    Record userRecord = dslContext.fetchOne(dslContext.selectFrom("UserData").where(USERDATA.ID.eq(id)))
+    return BCrypt.checkpw(password, userRecord.get("password") as String)
   }
 }
