+++ src/main/groovy/filter/AuthenticationFilter.groovy
@@ -1,5 +1,6 @@
 package filter
 
+import dao.UserDao
 import exception.InvalidTokenException
 import exception.UserNotExistsException
 import org.restlet.Request
@@ -12,19 +13,21 @@ import javax.inject.Inject
 class AuthenticationFilter extends Filter {
   private final UserDataValidation userDataValidation
   private final JwtUtil jwtUtil;
+  private final UserDao userDao
 
   @Inject
-  AuthenticationFilter(UserDataValidation userDataValidation, JwtUtil jwtUtil) {
+  AuthenticationFilter(UserDataValidation userDataValidation, JwtUtil jwtUtil, UserDao userDao) {
     this.userDataValidation = userDataValidation
     this.jwtUtil = jwtUtil
+    this.userDao = userDao
   }
 
   @Override
   protected int beforeHandle(Request request, Response response) {
     String token = request.getHeaders().getFirstValue("authorization")
-    if (!userDataValidation.isTokenValid(token)) {
+    if (!userDataValidation.isTokenPresent(token)) {
       throw new InvalidTokenException("Request header must have valid authentication token associated with it")
-    } else if (!userDataValidation.isUserExists(jwtUtil.extractUsername(token))) {
+    } else if (!userDao.isUserExists(jwtUtil.extractUsername(token))) {
       throw new UserNotExistsException("User not exists")
     } else if (!jwtUtil.validateToken(token, jwtUtil.extractUsername(token))) {
       throw new InvalidTokenException("Unauthorized request, log in and try again")
